Remove all the deprecated stuff from beta 16 and before
