+++ composer.json
@@ -74,8 +74,7 @@
             "Flarum\\": "src/"
         },
         "files": [
-            "src/helpers.php",
-            "src/TranslatorInterface.php"
+            "src/helpers.php"
         ]
     },
     "autoload-dev": {
