+++ src/Discussion/Search/DiscussionSearcher.php
@@ -10,11 +10,8 @@
 namespace Flarum\Discussion\Search;
 
 use Flarum\Discussion\DiscussionRepository;
-use Flarum\Discussion\Event\Searching;
-use Flarum\Query\QueryCriteria;
 use Flarum\Search\AbstractSearcher;
 use Flarum\Search\GambitManager;
-use Flarum\Search\SearchState;
 use Flarum\User\User;
 use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Database\Eloquent\Builder;
@@ -48,15 +45,5 @@ class DiscussionSearcher extends AbstractSearcher
     protected function getQuery(User $actor): Builder
     {
         return $this->discussions->query()->select('discussions.*')->whereVisibleTo($actor);
-    }
-
-    /**
-     * @deprecated along with the Searching event, remove in Beta 17.
-     */
-    protected function mutateSearch(SearchState $search, QueryCriteria $criteria)
-    {
-        parent::mutateSearch($search, $criteria);
-
-        $this->events->dispatch(new Searching($search, $criteria));
     }
 }
