--- src/Event/ConfigureUserGambits.php
@@ -1,17 +1,0 @@
-<?php
-
-/*
- * This file is part of Flarum.
- *
- * For detailed copyright and license information, please view the
- * LICENSE file that was distributed with this source code.
- */
-
-namespace Flarum\Event;
-
-/**
- * @deprecated beta 16, removed in beta 17
- */
-class ConfigureUserGambits extends AbstractConfigureGambits
-{
-}
