+++ src/Extend/ApiSerializer.php
@@ -75,28 +75,6 @@ class ApiSerializer implements ExtenderInterface
     }
 
     /**
-     * Add to or modify the attributes array of this serializer.
-     *
-     * @param callable|string $callback
-     *
-     * The callback can be a closure or an invokable class, and should accept:
-     * - $serializer: An instance of this serializer.
-     * - $model: An instance of the model being serialized.
-     * - $attributes: An array of existing attributes.
-     *
-     * The callable should return:
-     * - An array of additional attributes to merge with the existing array.
-     *   Or a modified $attributes array.
-     *
-     * @deprecated in beta 16, removed in beta 17
-     * @return self
-     */
-    public function mutate($callback)
-    {
-        return $this->attributes($callback);
-    }
-
-    /**
      * Establish a simple hasOne relationship from this serializer to another serializer.
      * This represents a one-to-one relationship.
      *
