+++ src/Foundation/AbstractServiceProvider.php
@@ -15,12 +15,6 @@ use Illuminate\Support\ServiceProvider;
 abstract class AbstractServiceProvider extends ServiceProvider
 {
     /**
-     * @deprecated beta 16, remove beta 17
-     * @var Container
-     */
-    protected $app;
-
-    /**
      * @var Container
      */
     protected $container;
@@ -30,7 +24,6 @@ abstract class AbstractServiceProvider extends ServiceProvider
      */
     public function __construct(Container $container)
     {
-        $this->app = $container;
         $this->container = $container;
     }
