+++ src/Http/Rememberer.php
@@ -32,35 +32,15 @@ class Rememberer
     /**
      * Sets the remember cookie on a response.
      * @param ResponseInterface $response
-     * @param RememberAccessToken $token The remember token to set on the response. Use of non-remember token is deprecated in beta 16, removed eta 17.
+     * @param RememberAccessToken $token The remember token to set on the response.
      * @return ResponseInterface
      */
-    public function remember(ResponseInterface $response, AccessToken $token)
+    public function remember(ResponseInterface $response, RememberAccessToken $token)
     {
-        if (! ($token instanceof RememberAccessToken)) {
-            trigger_error('Parameter $token of type AccessToken is deprecated in beta 16, must be instance of RememberAccessToken in beta 17', E_USER_DEPRECATED);
-
-            $token->type = 'session_remember';
-            $token->save();
-        }
-
         return FigResponseCookies::set(
             $response,
             $this->cookie->make(self::COOKIE_NAME, $token->token, RememberAccessToken::rememberCookieLifeTime())
         );
-    }
-
-    /**
-     * @param ResponseInterface $response
-     * @param $userId
-     * @return ResponseInterface
-     * @deprecated beta 16, removed beta 17. Use remember() with a token
-     */
-    public function rememberUser(ResponseInterface $response, $userId)
-    {
-        $token = RememberAccessToken::generate($userId);
-
-        return $this->remember($response, $token);
     }
 
     public function forget(ResponseInterface $response)
