+++ src/Http/SessionAuthenticator.php
@@ -15,20 +15,10 @@ class SessionAuthenticator
 {
     /**
      * @param Session $session
-     * @param AccessToken|int $token Token or user ID. Use of User ID is deprecated in beta 16, will be removed in beta 17
+     * @param AccessToken $token
      */
-    public function logIn(Session $session, $token)
+    public function logIn(Session $session, AccessToken $token)
     {
-        // Backwards compatibility with $userId as parameter
-        // Remove in beta 17
-        if (! ($token instanceof AccessToken)) {
-            $userId = $token;
-
-            trigger_error('Parameter $userId is deprecated in beta 16, will be replaced by $token in beta 17', E_USER_DEPRECATED);
-
-            $token = SessionAccessToken::generate($userId);
-        }
-
         $session->regenerate(true);
         $session->put('access_token', $token->token);
     }
