+++ src/Locale/Translator.php
@@ -12,9 +12,8 @@ namespace Flarum\Locale;
 use Illuminate\Contracts\Translation\Translator as TranslatorContract;
 use Symfony\Component\Translation\MessageCatalogueInterface;
 use Symfony\Component\Translation\Translator as BaseTranslator;
-use Symfony\Component\Translation\TranslatorInterface;  // Defined locally as BC layer
 
-class Translator extends BaseTranslator implements TranslatorContract, TranslatorInterface
+class Translator extends BaseTranslator implements TranslatorContract
 {
     const REFERENCE_REGEX = '/^=>\s*([a-z0-9_\-\.]+)$/i';
