--- src/Search/SearchResults.php
@@ -1,17 +1,0 @@
-<?php
-
-/*
- * This file is part of Flarum.
- *
- * For detailed copyright and license information, please view the
- * LICENSE file that was distributed with this source code.
- */
-
-namespace Flarum\Search;
-
-/**
- * @deprecated beta 16, remove beta 17
- */
-class SearchResults
-{
-}
