+++ tests/integration/api/posts/ListTest.php
@@ -116,44 +116,6 @@ class ListTests extends TestCase
 
     /**
      * @test
-     * @deprecated
-     */
-    public function user_filter_works()
-    {
-        $response = $this->send(
-            $this->request('GET', '/api/posts', ['authenticatedAs' => 1])
-                ->withQueryParams([
-                    'filter' => ['user' => 'admin'],
-                ])
-        );
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $data = json_decode($response->getBody()->getContents(), true);
-
-        $this->assertEquals(['1', '2'], Arr::pluck($data['data'], 'id'));
-    }
-
-    /**
-     * @test
-     * @deprecated
-     */
-    public function user_filter_works_with_multiple_values()
-    {
-        $response = $this->send(
-            $this->request('GET', '/api/posts', ['authenticatedAs' => 1])
-                ->withQueryParams([
-                    'filter' => ['user' => 'admin,normal'],
-                ])
-        );
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $data = json_decode($response->getBody()->getContents(), true);
-
-        $this->assertEquals(['1', '2', '3', '4', '5'], Arr::pluck($data['data'], 'id'));
-    }
-
-    /**
-     * @test
      */
     public function discussion_filter_works()
     {
@@ -240,25 +202,5 @@ class ListTests extends TestCase
         $data = json_decode($response->getBody()->getContents(), true);
 
         $this->assertEquals(['1', '3', '5'], Arr::pluck($data['data'], 'id'));
-    }
-
-    /**
-     * @deprecated beta 16, remove beta 17
-     * @test
-     */
-    public function deprecated_configure_posts_query_extension_still_works()
-    {
-        $this->app()->getContainer()->make('events')->listen(ConfigurePostsQuery::class, function (ConfigurePostsQuery $event) {
-            $event->query->where('id', '1');
-        });
-
-        $response = $this->send(
-            $this->request('GET', '/api/posts', ['authenticatedAs' => 1])
-        );
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $data = json_decode($response->getBody()->getContents(), true);
-
-        $this->assertEquals(['1'], Arr::pluck($data['data'], 'id'));
     }
 }
